feat: support for ManyToOne and OneToMany relations
In this commit the support for OneToMany and ManyToOne relationships are implemented.

First of all the foreign key is entered in the table.

Then if we find a OneToMany we will count how many items there are of that subitem (child) and if there are more than 0. We will query for them